Fix Treble Check
Some custom ROMs have the prop "ro.treble.enabled", but have it set to false. Currently, kernels fail to flash in TWRP because they fail the check, so fix it.

OmniROM Treskmod: 
OnePlus5:/ $ getprop ro.treble.enabled
false
OnePlus5:/ $

Florian in your group also couldn't flash it on LineageOS with this check in.

(I hope I did this right, apologize if not. This is my first proper PR.)